requirements.txt
attrs
pychromecast
